+++ tests/unit/content/extractors/NextJsExtractor.test.js
@@ -4,11 +4,9 @@
 
 import Logger from '../../../../scripts/utils/Logger.js';
 import { sanitizeUrlForLogging } from '../../../../scripts/utils/LogSanitizer.js';
-import * as NextJsDataResolver from '../../../../scripts/content/extractors/NextJsDataResolver.js';
 import { NextJsExtractor } from '../../../../scripts/content/extractors/NextJsExtractor.js';
 import { NEXTJS_CONFIG } from '../../../../scripts/config/shared/content.js';
 import yahooNewsRscFixture from '../../../fixtures/json/nextjs-rsc-yahoo-news.json';
-import bbcNewsBlocksFixture from '../../../fixtures/json/bbc-news-blocks.json';
 
 // Mock Logger to avoid cluttering test output
 jest.mock('../../../../scripts/utils/Logger.js', () => ({
@@ -938,654 +936,6 @@ describe('NextJsExtractor', () => {
       const result = NextJsExtractor._resolvePageOriginAndPath(mockEmptyDoc);
       expect(result.origin).toBe('http://localhost');
       expect(result.pathname).toBe('/');
-    });
-  });
-
-  describe('scoring helpers', () => {
-    describe('_scoreKeysDimension', () => {
-      it('同時包含合法的 title 和 author 時加 15 分', () => {
-        const node = { title: 'Some Title', author: 'Some Author' };
-        expect(NextJsDataResolver.scoreKeysDimension(node)).toBe(15);
-      });
-
-      it('僅包含合法的 title 時加 10 分', () => {
-        const node = { title: 'Some Title' };
-        expect(NextJsDataResolver.scoreKeysDimension(node)).toBe(10);
-      });
-
-      it('僅包含合法的 author 時加 5 分', () => {
-        const node = { author: 'Some Author' };
-        expect(NextJsDataResolver.scoreKeysDimension(node)).toBe(5);
-      });
-
-      it('空節點或無匹配特徵時得 0 分', () => {
-        const node = {};
-        expect(NextJsDataResolver.scoreKeysDimension(node)).toBe(0);
-      });
-
-      it('非物件節點時得 0 分', () => {
-        expect(NextJsDataResolver.scoreKeysDimension(null)).toBe(0);
-        expect(NextJsDataResolver.scoreKeysDimension(undefined)).toBe(0);
-      });
-    });
-
-    describe('_scoreStructuralDimension', () => {
-      it('包含非空 paragraphs 陣列時加 40 分', () => {
-        const node = { paragraphs: ['Para 1'] };
-        expect(NextJsDataResolver.scoreStructuralDimension(node)).toBe(40);
-      });
-
-      it('paragraphs 陣列為空或非陣列時得 0 分', () => {
-        expect(NextJsDataResolver.scoreStructuralDimension({ paragraphs: [] })).toBe(0);
-        expect(NextJsDataResolver.scoreStructuralDimension({})).toBe(0);
-      });
-
-      it('非物件節點時得 0 分', () => {
-        expect(NextJsDataResolver.scoreStructuralDimension(null)).toBe(0);
-        expect(NextJsDataResolver.scoreStructuralDimension(undefined)).toBe(0);
-      });
-    });
-
-    describe('_scoreContentDimension', () => {
-      it('包含 text (字串) 且有 id 時加 15 分', () => {
-        const node = { text: 'Some text', id: '123' };
-        expect(NextJsDataResolver.scoreContentDimension(node)).toBe(15);
-      });
-
-      it('包含 content (字串且長度 > 100) 時加 20 分', () => {
-        const longContent = 'A'.repeat(101);
-        const node = { content: longContent };
-        expect(NextJsDataResolver.scoreContentDimension(node)).toBe(20);
-      });
-
-      it('同時滿足 text+id 與 content (> 100) 時累加 35 分', () => {
-        const longContent = 'A'.repeat(101);
-        const node = { text: 'Some text', id: '123', content: longContent };
-        expect(NextJsDataResolver.scoreContentDimension(node)).toBe(35);
-      });
-
-      it('條件不滿足時得 0 分', () => {
-        expect(NextJsDataResolver.scoreContentDimension({ text: 'Some text' })).toBe(0); // 缺 id
-        expect(NextJsDataResolver.scoreContentDimension({ content: 'short content' })).toBe(0); // 長度 <= 100
-      });
-
-      it('非物件節點時得 0 分', () => {
-        expect(NextJsDataResolver.scoreContentDimension(null)).toBe(0);
-        expect(NextJsDataResolver.scoreContentDimension(undefined)).toBe(0);
-      });
-    });
-
-    describe('_scoreStructureAndText', () => {
-      it('應正確加總各個特徵維度的分數', () => {
-        const node = {
-          title: 'Title',
-          author: 'Author',
-          paragraphs: ['P1'],
-          text: 'Text',
-          id: '999',
-          content: 'C'.repeat(105),
-        };
-        // Keys (15) + Structural (40) + Content (35) = 90
-        expect(NextJsDataResolver.scoreStructureAndText(node)).toBe(90);
-      });
-    });
-  });
-
-  describe('BBC format (_isBbcFormat / _convertBbcBlocks / _extractBbcText)', () => {
-    // 模擬 BBC __NEXT_DATA__.props.pageProps.pageData.content.model 結構
-    const buildBbcNextData = blocks => ({
-      props: {
-        pageProps: {
-          pageData: {
-            content: {
-              model: { blocks },
-            },
-            promo: {
-              headlines: { seoHeadline: 'BBC Test Article' },
-            },
-          },
-        },
-      },
-    });
-
-    it('_isBbcFormat: 應識別 BBC {type, model} 格式', () => {
-      const bbcBlocks = [{ type: 'paragraph', model: { text: 'hello' } }];
-      expect(NextJsExtractor._isBbcFormat(bbcBlocks)).toBe(true);
-    });
-
-    it('_isBbcFormat: 首項為 null 時仍應識別後續 BBC block', () => {
-      const bbcBlocks = [null, { type: 'paragraph', model: { text: 'hello' } }];
-      expect(NextJsExtractor._isBbcFormat(bbcBlocks)).toBe(true);
-    });
-
-    it('_isBbcFormat: 應拒絕標準 {blockType, text} 格式', () => {
-      const standardBlocks = [{ blockType: 'paragraph', text: 'hello' }];
-      expect(NextJsExtractor._isBbcFormat(standardBlocks)).toBe(false);
-    });
-
-    it('應從 BBC 結構提取 heading + paragraph + image', () => {
-      const fixture = bbcNewsBlocksFixture;
-      const mockJson = buildBbcNextData(fixture.blocks);
-      mockDoc.querySelector.mockReturnValue({ textContent: JSON.stringify(mockJson) });
-      mockDoc.querySelectorAll.mockReturnValue([]);
-
-      const result = NextJsExtractor.extract(mockDoc);
-
-      expect(result).not.toBeNull();
-      // headline → heading_1, 2x text → 2 paragraphs, image → image
-      expect(result.blocks).toHaveLength(4);
-      expect(result.blocks[0].type).toBe('heading_1');
-      expect(result.blocks[0].heading_1.rich_text[0].text.content).toBe('文章標題');
-      expect(result.blocks[1].type).toBe('paragraph');
-      expect(result.blocks[1].paragraph.rich_text[0].text.content).toBe('第一段文字');
-      expect(result.blocks[2].type).toBe('paragraph');
-      expect(result.blocks[3].type).toBe('image');
-      expect(result.blocks[3].image.external.url).toBe(
-        'https://ichef.bbci.co.uk/ace/ws/1024/cpsprodpb/a985/live/test-image.jpg.webp'
-      );
-      expect(result.blocks[3].image.caption[0].text.content).toBe('圖片說明');
-      expect(result.metadata.title).toBe('BBC Test Article');
-    });
-
-    it('頂層 BBC blocks 首項為 null 時不應讓整體提取失敗', () => {
-      const bbcBlocks = [
-        null,
-        { type: 'headline', model: { blocks: [{ model: { text: '文章標題' } }] } },
-        {
-          type: 'text',
-          model: {
-            blocks: [{ type: 'paragraph', model: { text: '第一段文字', blocks: [] } }],
-          },
-        },
-        {
-          type: 'text',
-          model: {
-            blocks: [{ type: 'paragraph', model: { text: '第二段文字', blocks: [] } }],
-          },
-        },
-      ];
-
-      const mockJson = buildBbcNextData(bbcBlocks);
-      mockDoc.querySelector.mockReturnValue({ textContent: JSON.stringify(mockJson) });
-      mockDoc.querySelectorAll.mockReturnValue([]);
-
-      const result = NextJsExtractor.extract(mockDoc);
-
-      expect(result).not.toBeNull();
-      expect(result.metadata.title).toBe('BBC Test Article');
-      expect(result.blocks).toHaveLength(3);
-      expect(result.blocks[0].type).toBe('heading_1');
-    });
-
-    it('articleData.blocks 為空陣列時應回退使用 content.model.blocks 的 BBC 內容', () => {
-      const bbcBlocks = [
-        { type: 'headline', model: { blocks: [{ model: { text: '文章標題' } }] } },
-        {
-          type: 'text',
-          model: {
-            blocks: [{ type: 'paragraph', model: { text: '第一段文字', blocks: [] } }],
-          },
-        },
-        {
-          type: 'text',
-          model: {
-            blocks: [{ type: 'paragraph', model: { text: '第二段文字', blocks: [] } }],
-          },
-        },
-      ];
-
-      const mockJson = {
-        props: {
-          pageProps: {
-            pageData: {
-              blocks: [],
-              content: { model: { blocks: bbcBlocks } },
-              promo: { headlines: { seoHeadline: 'BBC Test Article' } },
-            },
-          },
-        },
-      };
-
-      mockDoc.querySelector.mockReturnValue({ textContent: JSON.stringify(mockJson) });
-      mockDoc.querySelectorAll.mockReturnValue([]);
-
-      const result = NextJsExtractor.extract(mockDoc);
-
-      expect(result).not.toBeNull();
-      expect(result.metadata.title).toBe('BBC Test Article');
-      expect(result.blocks).toHaveLength(3);
-      expect(result.blocks[0].type).toBe('heading_1');
-    });
-
-    it('byline/relatedContent blocks 應被跳過', () => {
-      const bbcBlocks = [
-        { type: 'headline', model: { blocks: [{ model: { text: '標題A' } }] } },
-        { type: 'byline', model: { blocks: [] } },
-        {
-          type: 'text',
-          model: { blocks: [{ type: 'paragraph', model: { text: 'Para 1', blocks: [] } }] },
-        },
-        { type: 'relatedContent', model: { blocks: [] } },
-        {
-          type: 'text',
-          model: { blocks: [{ type: 'paragraph', model: { text: 'Para 2', blocks: [] } }] },
-        },
-        {
-          type: 'text',
-          model: { blocks: [{ type: 'paragraph', model: { text: 'Para 3', blocks: [] } }] },
-        },
-      ];
-
-      const mockJson = buildBbcNextData(bbcBlocks);
-      mockDoc.querySelector.mockReturnValue({ textContent: JSON.stringify(mockJson) });
-      mockDoc.querySelectorAll.mockReturnValue([]);
-
-      const result = NextJsExtractor.extract(mockDoc);
-      expect(result).not.toBeNull();
-      // byline, relatedContent 各少一個 block
-      const types = result.blocks.map(b => b.type);
-      expect(types).not.toContain('byline');
-      expect(types).not.toContain('relatedContent');
-    });
-
-    it('品質門檻：BBC blocks < MIN_VALID_BLOCKS 時回傳 null', () => {
-      // 只有 1 個 byline（被跳過），結果 0 blocks → null
-      const bbcBlocks = [
-        { type: 'byline', model: { blocks: [] } },
-        { type: 'relatedContent', model: { blocks: [] } },
-      ];
-
-      const mockJson = buildBbcNextData(bbcBlocks);
-      mockDoc.querySelector.mockReturnValue({ textContent: JSON.stringify(mockJson) });
-      mockDoc.querySelectorAll.mockReturnValue([]);
-
-      const result = NextJsExtractor.extract(mockDoc);
-      // 轉換後 0 blocks < 3 → 應回傳 null
-      expect(result).toBeNull();
-    });
-
-    it('邊界測試：無效或空內容的處理 (分支覆蓋率)', () => {
-      const bbcBlocks = [
-        // 缺少 type 或 model 的 block
-        { type: 'paragraph' },
-        { model: { text: 'test' } },
-        // 空 headline
-        { type: 'headline', model: {} },
-        // 空 subheadline
-        { type: 'subheadline', model: { blocks: [] } },
-        // 即使是 text block 沒內容也不產生 blocks
-        { type: 'text', model: { blocks: [{ type: 'paragraph', model: {} }] } },
-        // 缺少圖片必要屬性的 image
-        { type: 'image', model: { blocks: [{ type: 'rawImage', model: {} }] } },
-        // 新增：涵蓋 _extractBbcText model !object (line 1140)
-        { type: 'headline', model: null },
-        // 新增：涵蓋 subheadline 正常提取 (line 1091)
-        {
-          type: 'subheadline',
-          model: { blocks: [{ type: 'paragraph', model: { text: 'Sub Heading Text' } }] },
-        },
-        // 新增：涵蓋 _extractBbcText child !object (line 1153)
-        {
-          type: 'text',
-          model: {
-            blocks: [null, 'string-is-not-object', { type: 'paragraph', model: { text: '' } }],
-          },
-        },
-        // 新增：涵蓋 switch-case 略過區塊的邏輯 (line 1110)
-        { type: 'byline', model: {} },
-        // 未知類型的 fallback block: 成功提取 fallback 文字
-        { type: 'unknown_type', model: { text: 'fallback_text' } },
-        // 未知類型的 fallback block: 提取不出文字
-        { type: 'unknown_empty', model: {} },
-      ];
-
-      // 只靠以上可能產出 1 個 block，不足品質門檻(3)，所以我們再加個有效片段
-      const validBlocks = [
-        { type: 'text', model: { blocks: [{ type: 'paragraph', model: { text: 'para 1' } }] } },
-        { type: 'text', model: { blocks: [{ type: 'paragraph', model: { text: 'para 2' } }] } },
-      ];
-
-      const mockJson = buildBbcNextData([...validBlocks, ...bbcBlocks]);
-      mockDoc.querySelector.mockReturnValue({ textContent: JSON.stringify(mockJson) });
-      mockDoc.querySelectorAll.mockReturnValue([]);
-
-      const result = NextJsExtractor.extract(mockDoc);
-      // 1 個來自 fallback_text, 1 個來自 subheadline, 2 個來自 para 1/2
-      expect(result).not.toBeNull();
-      expect(result.blocks).toHaveLength(4);
-      const texts = result.blocks.map(
-        b => b.paragraph?.rich_text[0]?.text?.content || b.heading_2?.rich_text[0]?.text?.content
-      );
-      expect(texts).toContain('fallback_text');
-      expect(texts).toContain('Sub Heading Text');
-      expect(texts).toContain('para 1');
-      expect(texts).toContain('para 2');
-    });
-  });
-
-  describe('convertBlocks', () => {
-    it('should return empty array for non-array input', () => {
-      expect(NextJsExtractor.convertBlocks(null)).toEqual([]);
-      expect(NextJsExtractor.convertBlocks(undefined)).toEqual([]);
-      expect(NextJsExtractor.convertBlocks({})).toEqual([]);
-    });
-
-    it('should return empty array for empty array input', () => {
-      expect(NextJsExtractor.convertBlocks([])).toEqual([]);
-    });
-
-    it('should skip malformed block items', () => {
-      const input = [null, undefined, 'bad-block', { blockType: 'paragraph', text: 'Valid' }];
-      const output = NextJsExtractor.convertBlocks(input);
-
-      expect(output).toHaveLength(1);
-      expect(output[0].type).toBe('paragraph');
-      expect(output[0].paragraph.rich_text[0].text.content).toBe('Valid');
-    });
-
-    it('should convert heading blocks', () => {
-      const input = [{ blockType: 'heading_1', text: 'H1' }];
-      const output = NextJsExtractor.convertBlocks(input);
-      expect(output[0].type).toBe('heading_1');
-      expect(output[0].heading_1.rich_text[0].text.content).toBe('H1');
-    });
-
-    it('should strip HTML from heading blocks', () => {
-      const input = [{ blockType: 'heading_1', text: '<h1>Title <i>Italic</i></h1>' }];
-      const output = NextJsExtractor.convertBlocks(input);
-      expect(output[0].type).toBe('heading_1');
-      expect(output[0].heading_1.rich_text[0].text.content).toBe('Title Italic');
-    });
-
-    it('should strip script tags from content', () => {
-      const input = [{ blockType: 'paragraph', text: '<script>alert("xss")</script>Hello' }];
-      const output = NextJsExtractor.convertBlocks(input);
-      expect(output[0].type).toBe('paragraph');
-      expect(output[0].paragraph.rich_text[0].text.content).toBe('Hello');
-    });
-
-    it('should strip style tags from content', () => {
-      const input = [{ blockType: 'paragraph', text: '<style>body { color: red; }</style>Hello' }];
-      const output = NextJsExtractor.convertBlocks(input);
-      expect(output[0].type).toBe('paragraph');
-      expect(output[0].paragraph.rich_text[0].text.content).toBe('Hello');
-    });
-
-    it('should strip HTML from quote blocks', () => {
-      const input = [{ blockType: 'quote', text: '<blockquote>Quote <b>Bold</b></blockquote>' }];
-      const output = NextJsExtractor.convertBlocks(input);
-      expect(output[0].type).toBe('quote');
-      expect(output[0].quote.rich_text[0].text.content).toBe('Quote Bold');
-    });
-
-    it('should strip HTML from paragraph text', () => {
-      const input = [{ blockType: 'paragraph', text: '<p>Hello <b>World</b></p>' }];
-      const output = NextJsExtractor.convertBlocks(input);
-      expect(output[0].type).toBe('paragraph');
-      expect(output[0].paragraph.rich_text[0].text.content).toBe('Hello World');
-    });
-
-    it('should filter out images without URL', () => {
-      const input = [
-        { blockType: 'image', image: { url: '' } }, // Empty URL
-        { blockType: 'image', image: { cdnUrl: 'https://valid.com' } },
-      ];
-      const output = NextJsExtractor.convertBlocks(input);
-      expect(output).toHaveLength(1);
-      expect(output[0].image.external.url).toBe('https://valid.com');
-    });
-
-    it('should convert HK01 summary blocks to quotes', () => {
-      const input = [
-        {
-          blockType: 'summary',
-          summary: ['Summary line 1', 'Summary line 2'],
-        },
-      ];
-      const output = NextJsExtractor.convertBlocks(input);
-      expect(output).toHaveLength(1);
-      expect(output[0].type).toBe('quote');
-      expect(output[0].quote.rich_text[0].text.content).toBe('Summary line 1\nSummary line 2');
-    });
-
-    it('should convert HK01 text blocks (htmlTokens) to paragraphs', () => {
-      const input = [
-        {
-          blockType: 'text',
-          htmlTokens: [
-            [
-              { type: 'text', content: 'Paragraph 1 part A, ' },
-              { type: 'text', content: 'part B.' },
-            ],
-            [{ type: 'text', content: 'Paragraph 2.' }],
-          ],
-        },
-      ];
-      const output = NextJsExtractor.convertBlocks(input);
-      expect(output).toHaveLength(2);
-      expect(output[0].type).toBe('paragraph');
-      expect(output[0].paragraph.rich_text[0].text.content).toBe('Paragraph 1 part A, part B.');
-      expect(output[1].type).toBe('paragraph');
-      expect(output[1].paragraph.rich_text[0].text.content).toBe('Paragraph 2.');
-    });
-
-    it('should convert HK01 complex tokens (boldLink, link) to paragraphs', () => {
-      const input = [
-        {
-          blockType: 'text',
-          htmlTokens: [
-            [
-              { type: 'text', content: 'Normal text ' },
-              { type: 'boldLink', content: 'Bold Link' },
-              { type: 'text', content: ' more text.' },
-            ],
-          ],
-        },
-      ];
-      const output = NextJsExtractor.convertBlocks(input);
-      expect(output).toHaveLength(1);
-      expect(output[0].type).toBe('paragraph');
-      expect(output[0].paragraph.rich_text[0].text.content).toBe(
-        'Normal text Bold Link more text.'
-      );
-    });
-
-    it('should convert list blocks (fallback to paragraph)', () => {
-      const input = [
-        { blockType: 'list', items: ['Item 1', 'Item 2'], ordered: false },
-        { blockType: 'list', items: ['First', 'Second'], ordered: true },
-      ];
-      const output = NextJsExtractor.convertBlocks(input);
-      expect(output).toHaveLength(4);
-      // First list (unordered)
-      expect(output[0].type).toBe('bulleted_list_item');
-      expect(output[0].bulleted_list_item.rich_text[0].text.content).toBe('Item 1');
-      expect(output[1].type).toBe('bulleted_list_item');
-      expect(output[1].bulleted_list_item.rich_text[0].text.content).toBe('Item 2');
-      // Second list (ordered)
-      expect(output[2].type).toBe('numbered_list_item');
-      expect(output[2].numbered_list_item.rich_text[0].text.content).toBe('First');
-      expect(output[3].type).toBe('numbered_list_item');
-      expect(output[3].numbered_list_item.rich_text[0].text.content).toBe('Second');
-    });
-
-    it('should convert code blocks (fallback to paragraph)', () => {
-      const input = [{ blockType: 'code', text: 'console.log("hello")', language: 'javascript' }];
-      const output = NextJsExtractor.convertBlocks(input);
-      // Currently falls back to paragraph as 'code' is not explicitly handled in switch
-      expect(output).toHaveLength(1);
-      expect(output[0].type).toBe('paragraph');
-      expect(output[0].paragraph.rich_text[0].text.content).toBe('console.log("hello")');
-    });
-  });
-
-  describe('_stripHtml', () => {
-    it('should handle null or undefined input', () => {
-      expect(NextJsExtractor._stripHtml(null)).toBe('');
-      expect(NextJsExtractor._stripHtml(undefined)).toBe('');
-    });
-
-    it('should remove script tags entirely', () => {
-      const html = '<div>Content<script>console.log("bad")</script></div>';
-      expect(NextJsExtractor._stripHtml(html)).toBe('Content');
-    });
-  });
-
-  describe('RSC payload parsing helpers', () => {
-    describe('_extractRscDataObject', () => {
-      it('should return null for non-array input', () => {
-        expect(NextJsDataResolver.extractRscDataObject(null)).toBeNull();
-        expect(NextJsDataResolver.extractRscDataObject({})).toBeNull();
-        expect(NextJsDataResolver.extractRscDataObject('not an array')).toBeNull();
-      });
-
-      it('should fallback to search loop and find object when array length is less than 4', () => {
-        const input = [{ target: 'found' }];
-        expect(NextJsDataResolver.extractRscDataObject(input)).toEqual({ target: 'found' });
-      });
-
-      it('should return index 3 item when length >= 4 and index 3 is a plain object', () => {
-        const input = ['$', '$L2a', null, { pageData: { atoms: [] } }];
-        expect(NextJsDataResolver.extractRscDataObject(input)).toEqual({
-          pageData: { atoms: [] },
-        });
-      });
-
-      it('should fallback to search loop when index 3 is null', () => {
-        const input = ['$', '$L2a', null, null, { pageData: { atoms: [] } }];
-        expect(NextJsDataResolver.extractRscDataObject(input)).toEqual({
-          pageData: { atoms: [] },
-        });
-      });
-
-      it('should return the first non-null plain object found in the array', () => {
-        const input = [null, 'string', { target: 'found' }, null, { second: 'ignored' }];
-        expect(NextJsDataResolver.extractRscDataObject(input)).toEqual({ target: 'found' });
-      });
-
-      it('should return null if no plain object is found', () => {
-        const input = [null, 'string', 123, true];
-        expect(NextJsDataResolver.extractRscDataObject(input)).toBeNull();
-      });
-
-      it('should skip nested arrays during fallback search', () => {
-        const input = [null, ['nested-array'], { target: 'found' }];
-        expect(NextJsDataResolver.extractRscDataObject(input)).toEqual({ target: 'found' });
-      });
-    });
-
-    describe('_tryParseRscLine', () => {
-      it('should return null when no colon is present', () => {
-        expect(NextJsDataResolver.tryParseRscLine('no colon here')).toBeNull();
-      });
-
-      it('should return null when payload does not start with { or [', () => {
-        expect(NextJsDataResolver.tryParseRscLine('1:invalid')).toBeNull();
-      });
-
-      it('should return inner object when payload is a valid RSC wrapper', () => {
-        const line = '4:["$", "$L2a", null, {"pageData": {"atoms": []}}]';
-        expect(NextJsDataResolver.tryParseRscLine(line)).toEqual({ pageData: { atoms: [] } });
-      });
-
-      it('should return the parsed object or array when it is not a classic RSC wrapper', () => {
-        const line = '4:[1, 2, 3]';
-        expect(NextJsDataResolver.tryParseRscLine(line)).toEqual([1, 2, 3]);
-
-        const lineWithObj = '4:[1, 2, {"target":"found"}]';
-        expect(NextJsDataResolver.tryParseRscLine(lineWithObj)).toEqual({ target: 'found' });
-      });
-
-      it('should return the parsed object when payload is a simple plain object JSON', () => {
-        const line = '3:{"someNoise": true}';
-        expect(NextJsDataResolver.tryParseRscLine(line)).toEqual({ someNoise: true });
-      });
-
-      it('should return null on malformed JSON instead of throwing error', () => {
-        const line = '3:{"malformed"';
-        expect(NextJsDataResolver.tryParseRscLine(line)).toBeNull();
-      });
-    });
-
-    describe('_fallbackParseRsc', () => {
-      it('should return null when chunk has no colon', () => {
-        expect(NextJsDataResolver.fallbackParseRsc('no colon')).toBeNull();
-      });
-
-      it('should return null on JSON parse failure', () => {
-        expect(NextJsDataResolver.fallbackParseRsc('1:{"bad"')).toBeNull();
-      });
-
-      it('should return inner object of a single-line RSC wrapper', () => {
-        const chunk = '4:["$", "$L2a", null, {"key": "val"}]';
-        expect(NextJsDataResolver.fallbackParseRsc(chunk)).toEqual({ key: 'val' });
-      });
-
-      it('should return parsed object when chunk is plain object', () => {
-        const chunk = '3:{"key": "val"}';
-        expect(NextJsDataResolver.fallbackParseRsc(chunk)).toEqual({ key: 'val' });
-      });
-
-      it('should return parsed object if _extractRscDataObject returns null but parsed is an object', () => {
-        const chunk = '1:{"foo": "bar"}';
-        expect(NextJsDataResolver.fallbackParseRsc(chunk)).toEqual({ foo: 'bar' });
-      });
-    });
-
-    describe('_parseMultiLineRsc', () => {
-      it('should return empty array for empty string', () => {
-        expect(NextJsDataResolver.parseMultiLineRsc('')).toEqual([]);
-      });
-
-      it('should parse single line rsc', () => {
-        const input = '3:{"someNoise":true}';
-        expect(NextJsDataResolver.parseMultiLineRsc(input)).toEqual([{ someNoise: true }]);
-      });
-
-      it('should filter out unparsable lines', () => {
-        const input = '3:{"someNoise":true}\n5:bad-line\n4:{"valid":true}';
-        expect(NextJsDataResolver.parseMultiLineRsc(input)).toEqual([
-          { someNoise: true },
-          { valid: true },
-        ]);
-      });
-
-      it('should return empty array when all lines are unparsable', () => {
-        const input = 'bad-line\nanother-bad';
-        expect(NextJsDataResolver.parseMultiLineRsc(input)).toEqual([]);
-      });
-    });
-
-    describe('_parseAppRouterScript', () => {
-      it('should return empty array for script content without push calls', () => {
-        expect(NextJsDataResolver.parseAppRouterScript('')).toEqual([]);
-        expect(NextJsDataResolver.parseAppRouterScript('console.log("hello")')).toEqual([]);
-      });
-
-      it('should parse script with single push call', () => {
-        const script = 'self.__next_f.push([1, "3:{\\\"someNoise\\\":true}\\n"])';
-        expect(NextJsDataResolver.parseAppRouterScript(script)).toEqual([{ someNoise: true }]);
-      });
-
-      it('should parse script with multiple push calls', () => {
-        const script =
-          'self.__next_f.push([1, "1:I[\\\"noise\\\"]\\n"])\nself.__next_f.push([1, "3:{\\\"valid\\\":true}\\n"])';
-        expect(NextJsDataResolver.parseAppRouterScript(script)).toEqual([
-          '1:I["noise"]\n',
-          { valid: true },
-        ]);
-      });
-
-      it('should skip malformed JSON inside push calls', () => {
-        const script =
-          'self.__next_f.push([1, "3:{\\\"valid\\\":true}\\n"])\nself.__next_f.push(invalid_json_here)';
-        expect(NextJsDataResolver.parseAppRouterScript(script)).toEqual([{ valid: true }]);
-      });
-
-      it('should skip pushing when closing parenthesis is missing', () => {
-        const script = 'self.__next_f.push([1, "3:{\\\"valid\\\":true}\\n"';
-        expect(NextJsDataResolver.parseAppRouterScript(script)).toEqual([]);
-      });
     });
   });
 });
